+++ Docs/Unit-test/README.md
@@ -19,11 +19,12 @@
 В этой папке располагается файл с именем `UnitTest.cs` и другие 
 [файлы](https://github.com/Students-of-the-city-of-Kostroma/Student-timetable/blob/dev/Docs/White-box/README.md), 
 предшествующие написанию модульных тестов. 
+Проект модульных тестов, должен повторять структуру папок. Физический путь файла и иерархия папок в проекте должны сходиться.
 
 ## Название тестовых объектов
 
 В файле `UnitTest.cs` должен содержать класс с именем `UT_ICTeacer`, 
-что означает что это тестовый комплект модульных тестов для тестирования сетода `Insert`, 
+что означает что это тестовый комплект модульных тестов для тестирования метода `Insert`, 
 контроллера `CTeacher`. Класс имеет пространство имен, соответсвующее пути файла, 
 в котором он расположен. Для класса `UT_ICTeacer` пространством имен будет 
 являться `UnitTestOfTimetableOfClasses.CTeacher.Insert`. Этот класс должен содержать
