+++ UnitTestOfTimetableOfClasses/CAcademicDegree/Update/UnitTest.cs
@@ -1,7 +1,7 @@
 ﻿using LibOfTimetableOfClasses;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CAcademicDegree.Update
 {
     [TestClass]
     public class UT_UCAcademicDegree
