+++ UnitTestOfTimetableOfClasses/CAuditor/Insert/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.Auditor.Insert
 {
     [TestClass]
     public class UT_ICAuditor
