+++ UnitTestOfTimetableOfClasses/CEnclosures/Insert/UnitTest.cs
@@ -1,7 +1,7 @@
 ﻿using LibOfTimetableOfClasses;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CEnclosures.Insert
 {
     [TestClass]
     public class UT_ICEnclosurse
