+++ UnitTestOfTimetableOfClasses/CGroup/Insert/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CGroup.Insert
 {
     [TestClass]
     public class UT_ICGroup
