+++ UnitTestOfTimetableOfClasses/CGroup/UPDATE/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CGroup.Update
 {
     [TestClass]
     public class UT_UCGroup
