+++ UnitTestOfTimetableOfClasses/CInstitute/Delete/UnitTest.cs
@@ -3,7 +3,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CInstitute.Delete
 {
     [TestClass]
     public class UT_DCInstitute
