+++ UnitTestOfTimetableOfClasses/CInstitute/Insert/UnitTest.cs
@@ -3,7 +3,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CInstitute.Insert
 {
     [TestClass]
     public class UT_ICInstitute
