+++ UnitTestOfTimetableOfClasses/CInstitute/Update/UnitTest.cs
@@ -1,14 +1,14 @@
 ﻿using LibOfTimetableOfClasses;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CInstitute.Update
 {
     [TestClass]
-    public class UT_Update_CInstitute
+    public class UT_UCInstitute
     {
         RefData refData = new RefData();
         [TestMethod]
-        public void Task_615_1() //замена всех атрибутов
+        public void UCInstitute_1() //замена всех атрибутов
         {
 
             MInstitute T_Institute = new MInstitute("Институт автоматизированных систем и технологий", "ИАСТ", "Лустгартен Ю.Л.", "Костромской Государственный Университет");
@@ -23,7 +23,7 @@ namespace UnitTestOfTimetableOfClasses
         }
 
         [TestMethod]
-        public void Task_615_2() //замена атрибутов Полное название
+        public void UCInstitute_2() //замена атрибутов Полное название
         {
 
             MInstitute T_Institute = new MInstitute("Институт автоматизированных систем и технологий", "ИАСТ", "Лустгартен Ю.Л.", "Костромской Государственный Университет");
@@ -40,7 +40,7 @@ namespace UnitTestOfTimetableOfClasses
         }
 
         [TestMethod]
-        public void Task_615_3() //краткая запись института уже есть в таблице
+        public void UCInstitute_3() //краткая запись института уже есть в таблице
         {
 
             MInstitute T_Institute = new MInstitute("Институт автоматизированных систем и технологий", "ИАСТ", "Лустгартен Ю.Л.", "Костромской Государственный Университет");
@@ -57,7 +57,7 @@ namespace UnitTestOfTimetableOfClasses
 
 
         [TestMethod]
-        public void Task_615_4() //директор уже есть в таблице
+        public void UCInstitute_4() //директор уже есть в таблице
         {
 
             MInstitute T_Institute = new MInstitute("Институт автоматизированных систем и технологий", "ИАСТ", "Лустгартен Ю.Л.", "Костромской Государственный Университет");
@@ -73,7 +73,7 @@ namespace UnitTestOfTimetableOfClasses
         }
 
         [TestMethod]
-        public void Task_615_5() //наименование ВУЗа уже есть в таблице
+        public void UCInstitute_5() //наименование ВУЗа уже есть в таблице
         {
 
             MInstitute T_Institute = new MInstitute("Институт автоматизированных систем и технологий", "ИАСТ", "Лустгартен Ю.Л.", "Костромской Государственный Университет");
