+++ UnitTestOfTimetableOfClasses/CTeacher/Delete/UnitTest.cs
@@ -1,7 +1,7 @@
 ﻿using LibOfTimetableOfClasses;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CTeacher.Delete
 {
     [TestClass]
     public class UT_DCTeacher
