+++ UnitTestOfTimetableOfClasses/CTeacher/Insert/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Data;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CTeacher.Insert
 {
     [TestClass]
     public class UT_ICTeacher
