+++ UnitTestOfTimetableOfClasses/CTeacher/Update/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CTeacher.Update
 {
     [TestClass]
     public class UT_UCTeacher
