+++ UnitTestOfTimetableOfClasses/CTitle/Delete/UnitTest.cs
@@ -2,7 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CTitle.Delete
 {
     [TestClass]
     public class UT_DCTitle
