+++ UnitTestOfTimetableOfClasses/CTypesOfOccupation/Update/UnitTest.cs
@@ -1,7 +1,7 @@
 ﻿using LibOfTimetableOfClasses;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace UnitTestOfTimetableOfClasses
+namespace UnitTestOfTimetableOfClasses.CTypesOfOccupation.Update
 {
     [TestClass]
     public class UT_UCTypesOfOccupation
